weibo2/requirements.txt
amqp==2.6.1
aniso8601==8.0.0
billiard==3.6.3.0
blinker==1.4
celery==4.4.7
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dnspython==1.16.0
dominate==2.5.1
email-validator==1.1.1
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Cache==0.13.1
Flask-Caching==1.9.0
Flask-Celery-Helper==1.1.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Moment==0.10.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gevent==20.6.2
greenlet==0.4.16
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
kombu==4.6.11
MarkupSafe==1.1.1
payjs==1.2.5
Pillow==7.2.0
PyMySQL==0.10.0
pytz==2020.1
redis==3.5.3
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.19
urllib3==1.25.10
vine==1.3.0
visitor==0.1.3
Werkzeug==0.15.6
WTForms==2.3.3
zipp==1.2.0
zope.event==4.4
zope.interface==5.1.0
